server/requirements.txt
fastapi==0.111.0
sqlalchemy[asyncio]==2.0.30
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.1
alembic==1.13.1
bcrypt==4.1.3
asyncpg==0.29.0
greenlet==3.0.3
python-dotenv==1.0.1